pyproject.toml
[tool.poetry]
name = "YVD Bot"
version = "1.0.0"
description = ""
authors = ["Cyero"]
license = "Open"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^2.25.1"
pytube = "^12.1.2"
psycopg2-binary = "^2.9.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
